requirements.txt
discord.py==2.6.3
PyNaCl>=1.5.0
python-dotenv==1.0.1
aiomysql==0.2.0
aiofiles==24.1.0
aiohttp==3.9.3
better-profanity==0.7.0
cryptography==44.0.3
crowdin-api-client==1.24.1
diskcache==5.6.3
pillow==11.2.1
pillow-avif-plugin==1.5.2
openai==1.99.9
urlextract==1.9.0
clean-text==0.6.0
apnggif==0.1.4
numpy==1.26.4
torch
torchvision
--extra-index-url https://download.pytorch.org/whl/cu126
transformers==4.48.3
pymilvus==2.5.14
Unidecode==1.4.0
opencv-python-headless==4.11.0.86
filetype==1.2.0
psutil==5.9.8
httpx==0.27.0
discord-ext-voice-recv==0.5.2a179
pynacl==1.5.0
faster-whisper==1.2.0
redis==6.4.0
